LayoutTests/editing/execCommand/query-font-size-expected.txt

test	html	queryCommandValue result
execCommand('FontSize', -2)	<font class="Apple-style-span" size="1">test</font>	1
execCommand('FontSize', -1)	<font class="Apple-style-span" size="2">test</font>	2
execCommand('FontSize', 0)	<font class="Apple-style-span" size="1">test</font>	1
execCommand('FontSize', 1)	<font class="Apple-style-span" size="1">test</font>	1
execCommand('FontSize', 2)	<font class="Apple-style-span" size="2">test</font>	2
execCommand('FontSize', 3)	test	3
execCommand('FontSize', 4)	<font class="Apple-style-span" size="4">test</font>	4
execCommand('FontSize', 5)	<font class="Apple-style-span" size="5">test</font>	5
execCommand('FontSize', 6)	<font class="Apple-style-span" size="6">test</font>	6
execCommand('FontSize', 7)	<font class="Apple-style-span" size="7">test</font>	7
execCommand('FontSize', '8px')	<font class="Apple-style-span" size="7">test</font>	7
execCommand('FontSize', '2px')	<font class="Apple-style-span" size="2">test</font>	2
manual CSS font-size: 3px	<span style="font-size: 3px">test</span>	1
manual CSS font-size: 0.2em	<span style="font-size: 0.2em">test</span>	1
manual CSS font-size: 17px	<span style="font-size: 17px">test</span>	4
manual CSS font-size: 31px	<span style="font-size: 31px">test</span>	6
manual CSS font-size: 50px	<span style="font-size: 50px">test</span>	7
manual CSS font-size: 5em	<span style="font-size: 5em">test</span>	7
manual CSS font-size: 10px	<span style="font-size: 10px">test</span>	1
monospace tests to ensure the bug 19161 does not affect queryCommandValue's values
execCommand('FontSize', -2)	<font class="Apple-style-span" size="1" face="monospace">test</font>	1
execCommand('FontSize', -1)	<font class="Apple-style-span" size="2" face="monospace">test</font>	2
execCommand('FontSize', 0)	<font class="Apple-style-span" size="1" face="monospace">test</font>	1
execCommand('FontSize', 1)	<font class="Apple-style-span" size="1" face="monospace">test</font>	1
execCommand('FontSize', 2)	<font class="Apple-style-span" size="2" face="monospace">test</font>	2
execCommand('FontSize', 3)	<font class="Apple-style-span" face="monospace">test</font>	3
execCommand('FontSize', 4)	<font class="Apple-style-span" size="4" face="monospace">test</font>	4
execCommand('FontSize', 5)	<font class="Apple-style-span" size="5" face="monospace">test</font>	5
execCommand('FontSize', 6)	<font class="Apple-style-span" size="6" face="monospace">test</font>	6
execCommand('FontSize', 7)	<font class="Apple-style-span" size="7" face="monospace">test</font>	7
